requirements.txt
psutil>=5.8.0
matplotlib>=3.5.0
pandas>=1.4.0
numpy>=1.21.0
seaborn>=0.11.0
plotly>=5.6.0
Flask>=2.0.0
PyYAML>=6.0
schedule>=1.1.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0